feat: move shared logic to base mob actor
