+++ src/modules/mobs/mob.ts
@@ -1,16 +1,24 @@
-import { Vector2, Vector3 } from "three";
+import { AStarFinder } from "astar-typescript";
+import { Color, Vector2, Vector3 } from "three";
 import { TBattleSide } from "../battlefield-container";
-import { Actor, ActorArgs, Container, Game } from "../game";
+import { Actor, ActorArgs, Composite, Container, Game } from "../game";
 import { HealthBar } from "../health-bar";
+import { HeadQuarters } from "../player";
+import { ParticleManager } from "../particle-systems";
+import { COLOR_PALETTE } from "../colors";
+import { PathfindingUtils } from "../pathfinding";
 
 export interface MobArgs extends ActorArgs {
   battleSide: TBattleSide;
   pos: Vector2;
   radius: number;
   health: number;
+  objective: HeadQuarters;
 }
 
 export class Mob extends Actor {
+  declare public mesh: Composite;
+
   public battleSide: TBattleSide;
 
   public pos: Vector2;
@@ -20,6 +28,8 @@ export class Mob extends Actor {
   public fullHealth: number;
   public health: number;
 
+  private objective: HeadQuarters;
+
   public afterSpawn(container: Container, pos: Vector2): void {
     super.afterSpawn(container, pos);
     this.hb.start(container);
@@ -28,6 +38,11 @@ export class Mob extends Actor {
   public beforeDeath(game: Game, container: Container, pos: Vector2): void {
     super.beforeDeath(game, container, pos);
     this.hb.end(container);
+
+    ParticleManager.createExplosion(container, this.mesh.position, [
+      new Color(COLOR_PALETTE.RED),
+      new Color(COLOR_PALETTE.DARK),
+    ]);
   }
 
   constructor(args: MobArgs) {
@@ -38,6 +53,7 @@ export class Mob extends Actor {
 
     this.fullHealth = args.health;
     this.health = args.health;
+    this.objective = args.objective;
 
     this.hb = new HealthBar({
       fullHealth: this.fullHealth,
@@ -57,8 +73,70 @@ export class Mob extends Actor {
   ): void {
     super.update(game, delta, container, pos);
 
-    this.hb.update(this.health, this.mesh.position.clone());
+    const ACCEPTABLE_DISTANCE = 0.01;
 
+    if (
+      this.pos.distanceTo(this.objective.position.clone()) < ACCEPTABLE_DISTANCE
+    ) {
+      this.objective.health -= this.objective.fullHealth * 0.1;
+      this.kill();
+
+      // Explosion
+      ParticleManager.createExplosion(
+        container,
+        this.mesh.position,
+        [
+          new Color(COLOR_PALETTE.RED),
+          new Color(COLOR_PALETTE.ORANGE),
+          new Color(COLOR_PALETTE.YELLOW),
+          new Color(COLOR_PALETTE.WHITE),
+        ],
+        30,
+        {
+          force: 3,
+          size: 0.8,
+        },
+      );
+    } else if (this.pos.distanceTo(pos) < ACCEPTABLE_DISTANCE) {
+      this.pos = pos;
+
+      const simpleGrid = PathfindingUtils.createSimpleGrid(
+        container.actorsGrid,
+      );
+      const pathFinder = new AStarFinder({
+        grid: {
+          matrix: simpleGrid,
+        },
+        diagonalAllowed: false,
+      });
+
+      const path = pathFinder.findPath(pos, this.objective.position.clone());
+
+      if (path.length > 1) {
+        const nextPos = new Vector2(path[1][0], path[1][1]);
+
+        container.actorsGrid[pos.x][pos.y].actors = container.actorsGrid[pos.x][
+          pos.y
+        ].actors.filter((a) => a !== this);
+        container.actorsGrid[nextPos.x][nextPos.y].actors.push(this);
+      }
+    } else {
+      const DELTA_MULTIPLIER = 0.0015;
+      const deltaMovement = delta * DELTA_MULTIPLIER;
+
+      const direction = pos.clone().sub(this.pos).normalize();
+      this.pos.add(direction.clone().multiplyScalar(deltaMovement));
+    }
+
+    this.hb.update(this.health, this.mesh.position.clone());
     if (this.health <= 0) this.kill();
+  }
+
+  public graphics(): void {
+    this.mesh.position = new Vector3(
+      this.pos.x,
+      this.mesh.position.y,
+      this.pos.y,
+    );
   }
 }
