+++ src/modules/mobs/walker.ts
@@ -1,13 +1,10 @@
-import { AStarFinder } from "astar-typescript";
 import { Color, Vector2, Vector3 } from "three";
 import { TBattleSide } from "../battlefield-container";
 import { COLOR_PALETTE } from "../colors";
 import { WOLRD_CONFIG } from "../config";
-import { Composite, Container, Game } from "../game";
+import { Composite } from "../game";
 import { MeshUtils } from "../mesh";
-import { PathfindingUtils } from "../pathfinding";
 import { HeadQuarters } from "../player";
-import { ParticleManager } from "../particle-systems";
 import { Mob } from "./mob";
 
 export interface WalkerArgs {
@@ -16,10 +13,6 @@ export interface WalkerArgs {
 }
 
 export class Walker extends Mob {
-  declare public mesh: Composite;
-
-  private objective: HeadQuarters;
-
   constructor(args: WalkerArgs) {
     const radius = WOLRD_CONFIG.TILE_SIZE / 4;
     const height = WOLRD_CONFIG.TILE_SIZE;
@@ -50,89 +43,7 @@ export class Walker extends Mob {
       pos: args.pos,
       radius: radius,
       health: 10,
+      objective: args.objective,
     });
-
-    this.objective = args.objective;
-  }
-
-  public update(
-    game: Game,
-    delta: number,
-    container: Container,
-    pos: Vector2,
-  ): void {
-    super.update(game, delta, container, pos);
-
-    const ACCEPTABLE_DISTANCE = 0.01;
-
-    if (
-      this.pos.distanceTo(this.objective.position.clone()) < ACCEPTABLE_DISTANCE
-    ) {
-      this.objective.health -= this.objective.fullHealth * 0.1;
-      this.kill();
-
-      // Explosion
-      ParticleManager.createExplosion(
-        container,
-        this.mesh.position,
-        [
-          new Color(COLOR_PALETTE.RED),
-          new Color(COLOR_PALETTE.ORANGE),
-          new Color(COLOR_PALETTE.YELLOW),
-          new Color(COLOR_PALETTE.WHITE),
-        ],
-        30,
-        {
-          force: 3,
-          size: 0.8,
-        },
-      );
-    } else if (this.pos.distanceTo(pos) < ACCEPTABLE_DISTANCE) {
-      this.pos = pos;
-
-      const simpleGrid = PathfindingUtils.createSimpleGrid(
-        container.actorsGrid,
-      );
-      const pathFinder = new AStarFinder({
-        grid: {
-          matrix: simpleGrid,
-        },
-        diagonalAllowed: false,
-      });
-
-      const path = pathFinder.findPath(pos, this.objective.position.clone());
-
-      if (path.length > 1) {
-        const nextPos = new Vector2(path[1][0], path[1][1]);
-
-        container.actorsGrid[pos.x][pos.y].actors = container.actorsGrid[pos.x][
-          pos.y
-        ].actors.filter((a) => a !== this);
-        container.actorsGrid[nextPos.x][nextPos.y].actors.push(this);
-      }
-    } else {
-      const DELTA_MULTIPLIER = 0.0015;
-      const deltaMovement = delta * DELTA_MULTIPLIER;
-
-      const direction = pos.clone().sub(this.pos).normalize();
-      this.pos.add(direction.clone().multiplyScalar(deltaMovement));
-    }
-  }
-
-  public graphics(): void {
-    this.mesh.position = new Vector3(
-      this.pos.x,
-      this.mesh.position.y,
-      this.pos.y,
-    );
-  }
-
-  public beforeDeath(game: Game, container: Container, pos: Vector2): void {
-    super.beforeDeath(game, container, pos);
-
-    ParticleManager.createExplosion(container, this.mesh.position, [
-      new Color(COLOR_PALETTE.RED),
-      new Color(COLOR_PALETTE.DARK),
-    ]);
   }
 }
